+++ src/date-picker/range-picker.jsx
@@ -84,8 +84,8 @@ export default class RangePicker extends Component {
          */
         footerRender: PropTypes.func,
         /**
-         * 日期范围值改变时的回调
-         * @return {MomentObject|String} 日期值
+         * 日期范围值改变时的回调 [ MomentObject|String, MomentObject|String ]
+         * @param {Array<MomentObject|String>} value 日期值
          */
         onChange: PropTypes.func,
         /**
@@ -132,7 +132,7 @@ export default class RangePicker extends Component {
          */
         popupTriggerType: PropTypes.oneOf(['click', 'hover']),
         /**
-         * 弹层对齐方式
+         * 弹层对齐方式, 具体含义见 OverLay文档
          */
         popupAlign: PropTypes.string,
         /**
@@ -484,13 +484,12 @@ export default class RangePicker extends Component {
 
     // 如果用户没有给定时间禁用逻辑，则给默认到禁用逻辑，即如果是同一天，则时间不能是同样的
     getDisabledTime = ({ startValue, endValue }) => {
-        const { disabledHours, disabledMinutes, disabledSeconds } = this.props;
+        const { disabledHours, disabledMinutes, disabledSeconds } = this.props.showTime || {};
 
         let disabledTime = {};
 
         if (startValue && endValue) {
             const isSameDay = startValue.format('L') === endValue.format('L');
-
             const newDisabledHours = isFunction(disabledHours) ?
                 disabledHours : (index) => {
                     if (isSameDay && index < startValue.hour()) {
@@ -605,6 +604,7 @@ export default class RangePicker extends Component {
 
         const datePanel = (<RangeCalendar
             showOtherMonth
+            format={this.format}
             defaultVisibleMonth={defaultVisibleMonth}
             onVisibleMonthChange={onVisibleMonthChange}
             disabledDate={disabledDate}