+++ src/date-picker/year-picker.jsx
@@ -55,7 +55,7 @@ class YearPicker extends Component {
         footerRender: PropTypes.func,
         /**
          * 日期值改变时的回调
-         * @return {MomentObject|String} 日期值
+         * @param {MomentObject|String} value 日期值
          */
         onChange: PropTypes.func,
         /**
@@ -89,7 +89,7 @@ class YearPicker extends Component {
          */
         popupTriggerType: PropTypes.oneOf(['click', 'hover']),
         /**
-         * 弹层对齐方式
+         * 弹层对齐方式, 具体含义见 OverLay文档
          */
         popupAlign: PropTypes.string,
         /**