+++ src/table/sticky/header.jsx
@@ -1,17 +1,30 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classnames from 'classnames';
 import Affix from '../../affix';
 
 /* eslint-disable react/prefer-stateless-function*/
 export default class StickHeader extends React.Component {
+    static propTypes = {
+        prefix: PropTypes.string,
+    }
     static contextTypes = {
         Header: PropTypes.any,
         offsetTop: PropTypes.number,
         affixProps: PropTypes.object
     }
     render() {
+        const { prefix } = this.props;
         const { Header, offsetTop, affixProps } = this.context;
-        return (<Affix {...affixProps} offsetTop={offsetTop}>
+
+        const newAffixProps = affixProps || {};
+        const { className, others } = newAffixProps;
+        const cls = classnames({
+            [`${prefix}table-affix`]: true,
+            className
+        });
+
+        return (<Affix {...others} className={cls} offsetTop={offsetTop}>
             <Header {...this.props}/>
         </Affix>);
     }